Refactor(naming): UseCase 동사 우선 네이밍으로 통일 및 Sync/Stats 클래스명 정리
