+++ semicolon/src/main/java/dukku/semicolon/boundedContext/product/app/cqrs/ProductSyncFacade.java
@@ -4,17 +4,15 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.UUID;
-
 @Component
 @RequiredArgsConstructor
 public class ProductSyncFacade {
-    private final ProductStatsUseCase productStatsUseCase;
+    private final SyncProductStatsUseCase syncProductStatsUseCase;
     private final DeleteProductSyncUseCase deleteProductSyncUseCase;
 
     @Transactional
     public void syncAllStats() {
-        productStatsUseCase.execute();
+        syncProductStatsUseCase.execute();
     }
 
     public void syncProductToElasticsearch(int productId) {
