+++ semicolon/src/main/java/dukku/semicolon/boundedContext/product/app/cqrs/SyncSearchProductStatsUseCase.java
@@ -17,7 +17,7 @@ import java.util.Map;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class SyncProductSearchStatsUseCase {
+public class SyncSearchProductStatsUseCase {
     private final ElasticsearchOperations elasticsearchOperations;
     private static final String INDEX_NAME = "products_v1";
