+++ semicolon/src/main/java/dukku/semicolon/boundedContext/product/in/ReviewStatsCleanupListener.java
@@ -1,6 +1,6 @@
 package dukku.semicolon.boundedContext.product.in;
 
-import dukku.semicolon.boundedContext.product.app.cqrs.SellerReviewStatsRedisSupport;
+import dukku.semicolon.boundedContext.product.app.cqrs.review.ReviewStatsRedisSupport;
 import dukku.semicolon.shared.product.event.ReviewStatsSyncedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -11,7 +11,7 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @RequiredArgsConstructor
 public class ReviewStatsCleanupListener {
 
-    private final SellerReviewStatsRedisSupport redisSupport;
+    private final ReviewStatsRedisSupport redisSupport;
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     public void handle(ReviewStatsSyncedEvent event) {
